Set site environment variables in local dev
We'll want more control over this in the future, especially in terms of
distinguishing between dev context and production context.

For now this is a starting point.

With this change any locally defined environment variable will also take
precendence over any env var set in the site settings or from an addon