Fix live tunnel port when proxyPort is taken
When starting the proxy we first check if proxyPort is free and otherwise we pick a random free port.

The live tunnel would always connect to the port specified in the settings, so in that case it would
start serving the wrong dev server on the .live domain.

This makes sure we pass on the new port to the live tunnel if changed